feat: upgrade function
Add metada as attribute named "meta". It is acessed by attr(data_body, which = "meta").

Split merge metadata in a new functio: metadata_join().

Add new function to count number of variables in the body data from excel file. This is usefull to check all files to be sure about dimensions and if have some exception.